requirements.txt
anyio==3.6.2
boto3==1.26.123
botocore==1.29.123
click==8.1.3
fastapi==0.95.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
loguru==0.7.0
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
